+++ inspect_wandb/weave/hooks.py
@@ -218,12 +218,13 @@ class WeaveEvaluationHooks(Hooks):
 
         if data.sample.scores is not None:
             for k,v in data.sample.scores.items():
-                score_metadata = (v.metadata or {}) | ({"explanation": v.explanation} if v.explanation is not None else {})
+                score_metadata = (v.metadata or {}) | ({"explanation": v.explanation} if v.explanation is not None else {}) | ({"answer": v.answer} if v.answer is not None else {})
                 with weave.attributes(score_metadata):
                     await sample_score_logger.alog_score(
                         scorer=k,
-                        score=format_score_types(v.value)
+                        score=format_score_types(v.value, scorer_name=k)
                     )
+
 
         # Total time
         if (